@@ -1,36 +1,36 @@
 <script lang="ts">
-	import type { ConvertedMediaItem } from '$lib/media';
-	import { formatMediaFileSize } from '$lib/mediaFileSizeFormatter';
+	import { type ConvertedMediaItem, formatMediaFileSize } from '$lib/media';
 	import { Copy, Download, Trash2 } from 'lucide-svelte';
 	import { copyMediaToClipboard } from '$lib/utils';
-	import { createEventDispatcher } from 'svelte';
 	import { deleteMediaItem } from '$lib/storage';
 	import { fade } from 'svelte/transition';
-	import { MediaType } from '$lib/files';
 	import { Button } from './ui/button';
 
-	export let item: ConvertedMediaItem;
+	const { item, previewImage }: { item: ConvertedMediaItem; previewImage: (url: string) => void } =
+		$props();
 
-	const dispatch = createEventDispatcher();
-
-	let objectURL = URL.createObjectURL(item.blob);
+	const objectURL = URL.createObjectURL(item.convertedFile);
 </script>
 
 <div
-	class="relative group border rounded-lg hover:shadow-md hover:-translate-y-1 transition-all"
+	class="group relative rounded-lg border transition-all hover:-translate-y-1 hover:shadow-md"
 	out:fade={{ duration: 300 }}
 >
 	<div
-		class="absolute flex justify-end p-2 w-full top-0 right-0 rounded-t-lg
-           bg-foreground/50 text-background opacity-0 pointer-events-none
-           group-hover:opacity-100 group-hover:pointer-events-auto
-           transition-all"
+		class="pointer-events-none absolute right-0 top-0 flex w-full justify-end rounded-t-lg
+           bg-foreground/75 p-2 text-background opacity-0
+           transition-all group-hover:pointer-events-auto
+           group-hover:opacity-100"
 	>
-		<Button on:click={() => copyMediaToClipboard(item.blob)} size="icon" variant="ghost">
+		<Button
+			onclick={async () => copyMediaToClipboard(item.convertedFile)}
+			size="icon"
+			variant="ghost"
+		>
 			<Copy class="size-4" />
 		</Button>
 		<Button
-			on:click={() => {
+			onclick={() => {
 				const link = document.createElement('a');
 				link.href = objectURL;
 				link.download = item.convertedName;
@@ -41,28 +41,28 @@
 		>
 			<Download class="size-4" />
 		</Button>
-		<Button on:click={() => deleteMediaItem(item.id)} size="icon" variant="ghost">
+		<Button onclick={() => deleteMediaItem(item.id)} size="icon" variant="ghost">
 			<Trash2 class="size-4 stroke-destructive" />
 		</Button>
 	</div>
 
-	{#if item.type === MediaType.Image}
-		<button class="w-full" on:click|preventDefault={() => dispatch('preview', objectURL)}>
+	{#if item.type === 'image'}
+		<button class="w-full" onclick={() => previewImage(objectURL)}>
 			<img
 				src={objectURL}
 				alt={item.originalName}
-				class="w-full h-44 object-cover mb-2 rounded-t-lg"
+				class="mb-2 h-44 w-full rounded-t-lg object-cover"
 			/>
 		</button>
-	{:else if item.type === MediaType.Video}
-		<!-- svelte-ignore a11y-media-has-caption -->
-		<video src={objectURL} class="w-full h-32 object-cover mb-2 rounded-t-lg" />
-	{:else if item.type === MediaType.Audio}
-		<audio src={objectURL} controls class="w-full mb-2" />
+	{:else if item.type === 'video'}
+		<!-- svelte-ignore a11y_media_has_caption -->
+		<video src={objectURL} class="mb-2 h-32 w-full rounded-t-lg object-cover"></video>
+	{:else if item.type === 'audio'}
+		<audio src={objectURL} controls class="mb-2 w-full"></audio>
 	{/if}
 
 	<div class="px-4 pb-2">
-		<p class="font-semibold truncate">{item.convertedName}</p>
+		<p class="truncate font-semibold">{item.convertedName}</p>
 		<p class="text-sm text-foreground/50">{formatMediaFileSize(item.size)}</p>
 	</div>
 </div>
