@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { AudioSettings, ImageSettings, VideoSettings } from '$lib/ffmpeg.d';
+	import type { AudioOptions, ImageOptions, VideoOptions } from '$lib/ffmpeg';
 	import {
 		baseAudioFormats,
 		baseImageFormats,
@@ -7,38 +7,45 @@
 		extendedAudioFormats,
 		extendedImageFormats,
 		extendedVideoFormats,
-		type MediaFormat
-	} from '$lib/mediaFormats';
+		isMediaFormat,
+		type MediaFormat,
+		type MediaType
+	} from '$lib/media';
 
 	import { writable, type Writable } from 'svelte/store';
 	import { Label } from '$components/ui/label';
 	import { Input } from '$components/ui/input';
-	import { MediaType } from '$lib/files';
 
 	import * as Select from '$components/ui/select';
 	import { onMount } from 'svelte';
 
 	const formatOptions = writable<{ label: string; value: MediaFormat }[]>([]);
 
-	export let advancedMode: Writable<boolean>;
-	export let mediaType: MediaType;
-	export let options: Writable<{
-		[MediaType.Audio]: AudioSettings;
-		[MediaType.Image]: ImageSettings;
-		[MediaType.Video]: VideoSettings;
-	}>;
+	const {
+		advancedMode,
+		mediaType,
+		options
+	}: {
+		advancedMode: Writable<boolean>;
+		mediaType: MediaType;
+		options: Writable<{
+			audio: AudioOptions;
+			image: ImageOptions;
+			video: VideoOptions;
+		}>;
+	} = $props();
 
 	onMount(updateFormatOptions);
 
 	function updateFormatOptions() {
 		switch (mediaType) {
-			case MediaType.Audio:
+			case 'audio':
 				formatOptions.set($advancedMode ? extendedAudioFormats : baseAudioFormats);
 				break;
-			case MediaType.Image:
+			case 'image':
 				formatOptions.set($advancedMode ? extendedImageFormats : baseImageFormats);
 				break;
-			case MediaType.Video:
+			case 'video':
 				formatOptions.set($advancedMode ? extendedVideoFormats : baseVideoFormats);
 				break;
 			default:
@@ -46,22 +53,30 @@
 				break;
 		}
 	}
+
+	// biome-ignore lint/style/useConst: <explanation>
+	let value = $state('');
+	const triggerContent = $derived(
+		$formatOptions.find((f) => f.value === value)?.label ?? 'Select a format'
+	);
 </script>
 
 <div class="space-y-4">
 	<div>
 		<Label for="format">Format</Label>
 		<Select.Root
+			type="single"
 			items={$formatOptions}
-			onSelectedChange={(state) => {
-				if (state) {
-					$options[mediaType].format = state.value;
-					console.debug('Selected format:', state.value);
+			onValueChange={(value) => {
+				if (value && isMediaFormat(value)) {
+					$options[mediaType].format = value;
+					console.debug('Selected format:', value);
 				}
 			}}
+			bind:value
 		>
 			<Select.Trigger>
-				<Select.Value placeholder="Select format" />
+				{triggerContent}
 			</Select.Trigger>
 			<Select.Content>
 				{#each $formatOptions as { label, value }}
@@ -72,10 +87,10 @@
 	</div>
 
 	{#if $advancedMode}
-		<div class="border-t border-foreground/20" />
+		<div class="border-t border-foreground/20"></div>
 		<div>
-			<h3 class="text-lg font-semibold mb-2">Advanced Options</h3>
-			{#if mediaType === MediaType.Audio}
+			<h3 class="mb-2 text-lg font-semibold">Advanced Options</h3>
+			{#if mediaType === 'audio'}
 				<div class="space-y-2">
 					<Label for="bitrate">Bitrate</Label>
 					<Input id="bitrate" bind:value={$options.audio.bitrate} placeholder="e.g., 128k" />
@@ -86,19 +101,12 @@
 					<Label for="channels">Channels</Label>
 					<Input id="channels" bind:value={$options.audio.channels} placeholder="e.g., 2" />
 				</div>
-			{:else if mediaType === MediaType.Image}
+			{:else if mediaType === 'image'}
 				<div class="space-y-2">
-					<Label for="resolution">Resolution</Label>
-					<Input
-						id="resolution"
-						bind:value={$options.image.resolution}
-						placeholder="e.g., 1280x720"
-					/>
-
 					<Label for="quality">Quality</Label>
 					<Input id="quality" bind:value={$options.image.quality} placeholder="e.g., 100" />
 				</div>
-			{:else if mediaType === MediaType.Video}
+			{:else if mediaType === 'video'}
 				<div class="space-y-2">
 					<Label for="resolution">Resolution</Label>
 					<Input
@@ -106,7 +114,6 @@
 						bind:value={$options.video.resolution}
 						placeholder="e.g., 1280x720"
 					/>
-
 					<Label for="bitrate">Bitrate</Label>
 					<Input id="bitrate" bind:value={$options.video.bitrate} placeholder="e.g., 1000k" />
 
