@@ -2,15 +2,15 @@
 	import { Dialog as DialogPrimitive } from "bits-ui";
 	import { cn } from "$components/utils.js";
 
-	type $$Props = DialogPrimitive.DescriptionProps;
-
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+	let {
+		ref = $bindable(null),
+		class: className,
+		...restProps
+	}: DialogPrimitive.DescriptionProps = $props();
 </script>
 
 <DialogPrimitive.Description
+	bind:ref
 	class={cn("text-muted-foreground text-sm", className)}
-	{...$$restProps}
->
-	<slot />
-</DialogPrimitive.Description>
+	{...restProps}
+/>
